Move internal/testsuite to testsuite
