+++ tests/asset-store/main_test.go
@@ -3,7 +3,7 @@ package main
 import (
 	"testing"
 
-	"github.com/kyma-project/kyma/tests/asset-store/internal/testsuite"
+	"github.com/kyma-project/kyma/tests/asset-store/testsuite"
 	"github.com/onsi/gomega"
 	"github.com/vrischmann/envconfig"
 	restclient "k8s.io/client-go/rest"
