+++ tests/asset-store/testsuite/asset.go
